requirements.txt
# requirements.txt
streamlit
google-cloud-bigquery
google-cloud-compute
google-auth
google-adk
pandas
google-api-core
google-genai
nest-asyncio
requests
google-cloud-aiplatform
httpx
mistralai[gcp]
litellm
pydantic
streamlit-mermaid
PyGithub
python-dotenv
newspaper3k
google-api-python-client